requirements.txt
pandas>=1.5.0
jsonschema>=4.17.0
genson>=1.2.2
pyyaml>=6.0
click>=8.1.0
tqdm>=4.64.0
python-dotenv>=1.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
selenium>=4.8.0
webdriver-manager>=3.8.0
pytest>=7.3.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0 